+++ web/apps/photos/src/components/Upload/UploadProgress/resultSection.tsx
@@ -2,8 +2,8 @@ import { UPLOAD_RESULT } from "@/new/photos/services/upload/types";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import { CaptionedText } from "components/CaptionedText";
 import ItemList from "components/ItemList";
-import UploadProgressContext from "contexts/uploadProgress";
 import { useContext } from "react";
+import UploadProgressContext from "./context";
 import {
     SectionInfo,
     UploadProgressSection,
