+++ web/apps/photos/src/components/Upload/UploadProgress/title.tsx
@@ -10,7 +10,7 @@ import { useContext } from "react";
 
 import UnfoldLessIcon from "@mui/icons-material/UnfoldLess";
 import UnfoldMoreIcon from "@mui/icons-material/UnfoldMore";
-import UploadProgressContext from "contexts/uploadProgress";
+import UploadProgressContext from "./context";
 
 const UploadProgressTitleText = ({ expanded }) => {
     return (
